+++ PSIdentityManagerUtils/Private/Get-IdentityManagerSessionToUse.ps1
@@ -2,7 +2,7 @@ function Get-IdentityManagerSessionToUse {
   [CmdletBinding()]
   Param (
     [parameter(Mandatory = $false, Position = 0, HelpMessage = 'The session to use')]
-    [VI.DB.Entities.ISession] $Session = $null
+    $Session = $null
   )
 
   Begin {
@@ -15,9 +15,15 @@ function Get-IdentityManagerSessionToUse {
       if ($null -eq $Session) {
         if ($Global:imsessions.Count -eq 1) {
           $sessionToUse = $Global:imsessions[$Global:imsessions.Keys[0]].Session
+        } else {
+          throw [System.InvalidOperationException] 'There is more than one session active. You must specify which session should be used.'
         }
       } else {
         $sessionToUse = $Session
+      }
+
+      if ($null -eq $sessionToUse) {
+        throw [System.InvalidOperationException] 'There is no session. You must open a new session with New-IdentityManagerSession before proceeding.'
       }
 
       return $sessionToUse