+++ PSIdentityManagerUtils/Public/Get-EntityMethod.ps1
@@ -1,8 +1,16 @@
 ﻿function Get-EntityMethod() {
   Param (
     [parameter(Mandatory = $false, HelpMessage = 'The session to use')]
-    [VI.DB.Entities.ISession] $Session = $null,
-    [parameter(Mandatory = $false, ValueFromPipeline=$true, HelpMessage = 'Entity to interact with')]
+    [ValidateScript({
+      try {
+        $_.GetType().ImplementedInterfaces.Contains([type]'VI.DB.Entities.ISession')
+      }
+      catch [System.Management.Automation.PSInvalidCastException] {
+        throw [System.Management.Automation.PSInvalidCastException] 'The given value is not a valid session.'
+      }
+    })]
+    $Session = $null,
+    [parameter(Mandatory = $false, Position = 0, ValueFromPipeline = $true, HelpMessage = 'Entity to interact with')]
     [VI.DB.Entities.IEntity] $Entity,
     [parameter(Mandatory = $false, HelpMessage = 'The tablename of the object')]
     [string] $Type,
@@ -20,6 +28,8 @@
     } catch {
       Resolve-Exception -ExceptionObject $PSitem
     }
+
+    $entityMethods = New-Object Collections.Generic.List[EntityMethod]
   }
 
   Process {
@@ -30,20 +40,19 @@
 
       $objectMethods = [VI.DB.Entities.Entity]::GetEntityMethodsAsync($Entity, $sessionToUse, $null, $noneToken).GetAwaiter().GetResult()
 
-      Write-Output 'Object methods'
       ForEach ($om in $objectMethods) {
-        Write-Output "`t" $om.Caption.Original
+        $entityMethods.Add([EntityMethod]::new('Object', $om.Name, $om.Caption.Original))
       }
 
       $customizerMethods = [VI.DB.Entities.Entity]::GetMethodsAsync($Entity, $sessionToUse, $noneToken).GetAwaiter().GetResult()
-      Write-Output 'Customizer methods'
       ForEach ($cm in $customizerMethods) {
-        Write-Output "`t" $cm.Key
+        $entityMethods.Add([EntityMethod]::new('Customizer', $cm.Key, $cm.Name))
       }
     } catch {
       Resolve-Exception -ExceptionObject $PSitem
     }
 
+    $entityMethods
   }
 
   End {