+++ PSIdentityManagerUtils/Public/Get-ImEvent.ps1
@@ -1,8 +1,16 @@
 ﻿function Get-ImEvent() {
   Param (
     [parameter(Mandatory = $false, HelpMessage = 'The session to use')]
-    [VI.DB.Entities.ISession] $Session = $null,
-    [parameter(Mandatory = $false, ValueFromPipeline=$true, HelpMessage = 'Entity to interact with')]
+    [ValidateScript({
+      try {
+        $_.GetType().ImplementedInterfaces.Contains([type]'VI.DB.Entities.ISession')
+      }
+      catch [System.Management.Automation.PSInvalidCastException] {
+        throw [System.Management.Automation.PSInvalidCastException] 'The given value is not a valid session.'
+      }
+    })]
+    $Session = $null,
+    [parameter(Mandatory = $false, Position = 0, ValueFromPipeline = $true, HelpMessage = 'Entity to interact with')]
     [VI.DB.Entities.IEntity] $Entity,
     [parameter(Mandatory = $false, HelpMessage = 'The tablename of the object')]
     [string] $Type,
@@ -41,7 +49,7 @@
       $entityCollection = $src.GetCollectionAsync($query, [VI.DB.Entities.EntityCollectionLoadType]::Slim, $noneToken).GetAwaiter().GetResult()
 
       ForEach ($e in $entityCollection) {
-        Write-Output (Get-EntityColumnValue -Entity $e -Column 'EventName')
+        Get-EntityColumnValue -Entity $e -Column 'EventName'
       }
     } catch {
       Resolve-Exception -ExceptionObject $PSitem