+++ PSIdentityManagerUtils/Public/New-Entity.ps1
@@ -2,7 +2,15 @@ function New-Entity {
   [CmdletBinding()]
   Param (
     [parameter(Mandatory = $false, HelpMessage = 'The session to use')]
-    [VI.DB.Entities.ISession] $Session = $null,
+    [ValidateScript({
+      try {
+        $_.GetType().ImplementedInterfaces.Contains([type]'VI.DB.Entities.ISession')
+      }
+      catch [System.Management.Automation.PSInvalidCastException] {
+        throw [System.Management.Automation.PSInvalidCastException] 'The given value is not a valid session.'
+      }
+    })]
+    $Session = $null,
     [parameter(Mandatory = $true, HelpMessage = 'The tablename of the object to create')]
     [ValidateNotNullOrEmpty()]
     [String] $Type,