+++ PSIdentityManagerUtils/Public/Set-Entity.ps1
@@ -2,8 +2,16 @@ function Set-Entity {
   [CmdletBinding()]
   Param (
     [parameter(Mandatory = $false, HelpMessage = 'The session to use')]
-    [VI.DB.Entities.ISession] $Session = $null,
-    [parameter(Mandatory = $false, ValueFromPipeline=$true, HelpMessage = 'Entity to interact with')]
+    [ValidateScript({
+      try {
+        $_.GetType().ImplementedInterfaces.Contains([type]'VI.DB.Entities.ISession')
+      }
+      catch [System.Management.Automation.PSInvalidCastException] {
+        throw [System.Management.Automation.PSInvalidCastException] 'The given value is not a valid session.'
+      }
+    })]
+    $Session = $null,
+    [parameter(Mandatory = $false, ValueFromPipeline = $true, HelpMessage = 'Entity to interact with')]
     [VI.DB.Entities.IEntity] $Entity = $null,
     [parameter(Mandatory = $false, HelpMessage = 'The tablename of the object to modify')]
     [string] $Type,